+++ realtime/util/adcpin.h
@@ -34,7 +34,7 @@ public:
     inline unsigned int channel() const { return _channel; }
     inline void setChannel(unsigned int channel) { _channel = channel; }
 
-    double readValue() const;
+    int32_t readValue() const;
 
 private:
     void changeMode();
